+++ zppy/mpas_analysis.py
@@ -73,13 +73,14 @@ def mpas_analysis(config, scriptDir):
             with open(scriptFile, 'w') as f:
                 f.write(template.render( **c ))
 
-            # Submit job
-            jobid = submitScript(scriptFile, dependFiles=dependencies)
-
-            # Update status file
-            with open(statusFile, 'w') as f:
-                f.write('WAITING %d\n' % (jobid))
-
-            # Add to the dependency list
-            dependencies.append(statusFile)
+            if not c['dry_run']:
+                # Submit job
+                jobid = submitScript(scriptFile, dependFiles=dependencies)
+                
+                # Update status file
+                with open(statusFile, 'w') as f:
+                    f.write('WAITING %d\n' % (jobid))
+                    
+                    # Add to the dependency list
+                    dependencies.append(statusFile)
